Separate each classes
